simple_lmk: Introduce simple low memory killer for Android
This is a simple replacement for the Android Low Memory Killer driver that
simply targets a predefined number of pages to be freed every time a memory
allocation attempt is about to utilize direct reclaim. Processes are killed
according to the priorities that Android gives them; in other words, the
least important processes are always killed first.

Simple LMK memory reclaim is triggered in the page-alloc slow path after
compaction has failed. When the Simple LMK memory reclaim process is
entered, memory is reclaimed in an infinite loop until the requested memory
allocation is fulfilled when pages are freed. OOM'd memory allocations are
placed on a queue and are prioritized over non-OOM'd memory allocations
when new pages are freed. This is to ensure that each OOM'd memory
allocation gets fulfilled in a timely manner.

Change-Id: I93e50abb6aab07e6a6a8b7358904fea3cdc3855e
Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>